feat(storage): replace rollout cache with bounded previews
Route preview publication through the authoritative artifact transaction, keep checkpoint retention configurable, and remove the persistent rollout cache path. Extend single/DDP/resume coverage and preserve SD3 timestep and policy dtype correctness.

+++ tests/native_parity/run_flow_grpo_sd3.py
@@ -1249,7 +1249,7 @@ def _rollout_native(
     )
     timesteps = (
         torch.as_tensor(adapter.pipeline.scheduler.timesteps)
-        .to(device=old_log_probs.device, dtype=torch.int64)
+        .to(device=old_log_probs.device)
         .reshape(1, transition_count)
         .expand(batch_size, transition_count)
         .clone()
